Several optimizations:
(1) CombinedSuffleSink initialized with 64 map slots to avoid double size operations
(2) The prediction loop is blocked to better fit the cache size
(3) In the tree node structure, we changed integer type to short type to make it smaller.

It seems (1) and (3) can optimize the performance by a few seconds respectively. Didn't see much improvement brought by (2).

+++ src/conf/headers/DataTypes.h
@@ -99,10 +99,10 @@ enum struct ModelType
 
 typedef struct {
 
-         int indexID;
+         short indexID;
          bool isLeaf;
-         int leftChild;
-         int rightChild;
+         short leftChild;
+         short rightChild;
          // returnClass will be the vaule to compare while this is not a leaf node
          float returnClass;
     } Node;
